Disconnect expired clients on release
After moving an app to use Tornadis, we started seeing connections leak.
We were using a pool of a limited size with a connection timeout. Since
the time delats are on connection state changes are calculated in
seconds, we tracked the connection leak down to
``ClientPool.release_client``. We found it possible for a client's
connection to expire while the client is in use most likely due to
rounding to whole seconds when calcuating a client's freshness. Since
the client is never appended back to the pool allowing it to be
disconnected
[here](https://github.com/thefab/tornadis/blob/master/tornadis/pool.py#L62).

This change checks if the client is still connected if it's not placed
back into the pool and disconnects it.